+++ app/Http/Controllers/MapController.php
@@ -3,30 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Traits\DealsWithObservationPermissions;
-use App\Http\Controllers\Traits\MapQuery;
 use App\Http\Controllers\Traits\Observes;
 use App\Http\Controllers\Traits\Responds;
 use App\Observation;
+use App\User;
 use Illuminate\Http\Request;
 use Cache;
 
 class MapController extends Controller
 {
-    use Responds, Observes, DealsWithObservationPermissions, MapQuery;
-
-    public function simpleMap(Request $request)
-    {
-        $this->validate($request, [
-            'bounds' => 'nullable|json',
-        ]);
-
-        $user = $request->user();
-        $bounds = json_decode($request->bounds);
-
-        $observations = $this->queryObservations($request->all(), $user, $bounds);
-
-        return $observations->get();
-    }
+    use Responds, Observes, DealsWithObservationPermissions;
 
     /**
      * Load observations for the map.
@@ -44,25 +30,11 @@ class MapController extends Controller
             'selectedFilter' => 'nullable|integer',
             'selectedConfirmation' => 'nullable|integer',
         ]);
-        info($request);
         $user = $request->user();
         $bounds = json_decode($request->bounds);
         $observations = $this->queryObservations($request->all(), $user, $bounds);
 
         return $this->success($observations->get());
-//        info($request->selectedCategories);
-//        info($request->selectedCollection);
-//        info($request->selectedFilter);
-//        info($request->selectedConfirmation);
-        $user = $request->user();
-        $isAdmin = false;
-        if ($user) {
-            $isAdmin = $user->isScientist() || $user->isAdmin();
-        }
-
-        $observations = $this->getObservationsFromDB($user, $isAdmin, $request->bounds);
-
-        return $this->success($observations);
     }
 
     /**
@@ -91,92 +63,12 @@ class MapController extends Controller
     }
 
     /**
-     * Query the database for observations.
-     *
-     * @param $user
-     * @param $isAdmin
-     * @param $bounds
-     * @return array
-     */
-    public function getObservationsFromDB($user, $isAdmin, $bounds): array
-    {
-        $observations = Observation::select([
-            'observations.id',
-            'observations.user_id',
-            'observation_category',
-//            'observations.images',
-            'observations.thumbnail',
-            'observations.longitude',
-            'observations.latitude',
-            'observations.data',
-            'observations.collection_date',
-//            'observations.location_accuracy',
-            'observations.is_private',
-//            'observations.address',
-            'observations.fuzzy_coords',
-            'observations.mobile_id',
-//            'observations.latin_name_id',
-//            'observations.has_private_comments',
-            'observations.custom_id',
-        ])->with('user', function ($query) {
-            $query->select([
-                'id',
-                'name',
-                'is_anonymous'
-            ]);
-        })->withCount([
-            'confirmations' => function ($query) {
-                $query->where('correct', true);
-            },
-        ]);
-
-        if ($bounds) {
-            $bounds = json_decode($bounds);
-            $observations->bounds($bounds);
-        }
-
-        if (!$user) {
-            $observations = $observations->where('is_private', false);
-        } elseif (!$isAdmin) {
-            $observations = $this->addPrivacyClause($observations, $user);
-        }
-
-        $observations = $observations->orderBy('observations.id', 'desc')->get();
-
-        if ($user) {
-            $collections = $user->collections->map(function ($collection) {
-                return $collection->id;
-            });
-
-            $observations->load([
-                'flags' => function ($query) use ($user) {
-                    $query->select([
-                        'id',
-                        'user_id',
-                        'observation_id'
-                    ]);
-                    $query->where('user_id', $user->id);
-                },
-                'collections' => function ($query) use ($collections) {
-                    $query->select([
-                        'id',
-                        'user_id',
-                    ]);
-                    $query->whereIn('collections.id', $collections);
-                },
-            ]);
-        }
-
-        return $this->prepForMap($observations, $isAdmin, $user);
-    }
-
-    /**
-     * Loads data for a single observation.
+     * Loads data for a selected observation within the map.
      * @param Request $request
      * @param Observation $observation
      * @return array
      */
-    public function getObservationData(Request $request, Observation $observation): array
+    public function showObservation(Request $request, Observation $observation): array
     {
         $user = $request->user();
         $isAdmin = false;
@@ -244,5 +136,66 @@ class MapController extends Controller
         ];
 
         return $data;
+    }
+
+    /**
+     * Loads data for all observations visible within bounds of the map.
+     * @param array $parameters
+     * @param User $user
+     * @param $bounds
+     * @return mixed
+     */
+    function queryObservations(array $parameters, User $user, $bounds)
+    {
+        $isAdmin = $user->isAdmin();
+        $isScientist = $user->isScientist();
+        $friends = $user->friends() + [$user->id];
+
+        $observations = Observation::query()
+            ->bounds($bounds)
+            ->when($isAdmin || $isScientist, function ($query) {
+                // Admin logic
+                $query->select(['id', 'latitude', 'longitude']);
+            }, function ($query) use ($friends) {
+                // Non-admin logic
+                // If the user is in the same group and is sharing then get latitude and longitude
+                // If the user owns the observation, same as above
+                // Otherwise, get fuzzy coords
+                $query->selectRaw('id, IF(user_id in (?), latitude, null) as latitude, IF(user_id in (?), longitude, null) as longitude, fuzzy_coords',
+                    [
+                        $friends,
+                        $friends,
+                    ]);
+            })
+            ->addSelect(['thumbnail', 'observation_category as title'])
+            ->when(!$user, function ($query) {
+                $query->where('is_private', false);
+            })
+            ->when(!empty($parameters['searchTerm']), function ($query) use ($parameters) {
+                $query->where(function ($query) use ($parameters) {
+                    $query->whereHas('user', function ($query) use ($parameters) {
+                        $query->where('users.name', 'like', "%{$parameters['searchTerm']}%");
+                    });
+                    $query->orWhere('observation_category', 'like', "%{$parameters['searchTerm']}%");
+                    if (isset($observation->data['otherLabel'])) {
+                        $query->orWhereJsonContains('otherLabel', ["%{$parameters['searchTerm']}%"]);
+                    }
+                    $query->orWhere('custom_id', 'like', "%{$parameters['searchTerm']}%");
+                    $query->orWhere('mobile_id', 'like', "%{$parameters['searchTerm']}%");
+                });
+            })
+            ->when(!empty($parameters['selectedCategories']), function ($query) use ($parameters) {
+                $query->whereIn('observation_category', $parameters['selectedCategories']);
+            })
+            ->when(!empty($parameters['selectedCollection']), function ($query) use ($parameters) {
+                $query->whereHas('collections', function ($query) use ($parameters) {
+                    $query->where('id', $parameters['selectedCollection']);
+                });
+            })
+            ->when(!empty($parameters['selectedConfirmation']), function ($query) use ($parameters) {
+                $query->whereHas('confirmations');
+            });
+
+        return $observations->orderBy('observations.id', 'desc');
     }
 }
